backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
python-socketio>=5.10.0
pydantic>=2.4.0
pydantic-settings>=2.0.3
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.26.0
python-multipart>=0.0.6
